+++ pyproject.toml
@@ -17,6 +17,7 @@ python-dotenv = "^0.17.1"
 typer = "^0.3.2"
 
 [tool.poetry.dev-dependencies]
+pylint = "^2.8.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
